requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Automat==0.8.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
Click==7.0
colorama==0.4.1
constantly==15.1.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Script==2.0.6
Flask-Twisted==0.1.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.2.0
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
mysqlclient==1.4.6
observable==1.0.3
packaging==19.2
pluggy==0.13.1
protobuf==3.11.0
py==1.8.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.4.3
pyparsing==2.4.5
pytest==5.3.1
six==1.13.0
SQLAlchemy==1.3.11
Twisted==19.10.0
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
zope.interface==4.7.1
